api/requirements.txt
spotipy
Flask
mutagen
requests
Pillow
dropbox
Flask-CORS
Flask-Limiter
Flask-WTF